audio: fix the Asphalt8 audio underrun issue when background [1/1]
PD#171292

more and more games are using opensles for game sound output
for lower latency, also they will limit the hal buffer latency
when send data. currently, the audio hal buffer size which
return to audiofinger is 8192. it is a little big which the
mixer thread goes to normal thread, not fastmixer thread.
the game sound will underrun always when BG.

Change-Id: I1c1f00c4d46e26486102b480a8f277fb9e41e8c1