audio: fix out of range access for dsp processing and aec
BUG=118157035

There are two issues related to the distorted output
for the NTS tests. The problem happens when NTS test
uses Dolby audio track in tunnel mode. In such cases,
the HAL layer will receive decoded PCM samples in 768
frames instead of 1024 frames for normal AF output based
on HAL reported size.

The first issue is the processing size of harman dsp
function. It uses the max buffer allocation size to
run the processing, but the allocation size may be bigger
than real numbers. If there is any normal audio streams
played (such as system sound) runs before the NTS test,
then the buffer will be created for 1024 frame size, and
the DSP processing unit is set to 1024. When we start
to run NTS, for Dolby audio track, the frame size is
768 and Harman DSP processing is still running in 1024
size, which makes wrong input into the internal filtering
calculation and that caused the "modulated" output.

The second issue is a potential crash, which happens
when timestamp is inserted into the speaker ring buffer
(aec_buf->spk_ring_buf). The original code seems have
an assumption that the input frame size is bigger than
the processing unit size (spk_buf_size 24576 bytes), so
it splits each copy to first set and second set. But
actually for Dolby case, each timestamp insertion will
only have 6144 (=768 * 8) size. So the copy will have
the chance to access out of malloced buffer and may cause
crash. This issue is fixed also.

Change-Id: I167111bf5501bd8cb451fdad427a55dffea88f9b
Signed-off-by: Tim Yao <tim.yao@amlogic.com>