Fix various checkbuild-related HAL issues
Mainly unused and undeclared variables.

One transitive dependency that needed to be explicit.

One unwinding of original code that tried to solve a name collision with
an existing function taking a different kind of pointer, by pretending
to use the same type in the header. The solution was to just remove an
include line to avoid the naming conflict happening in the first place.

Test: lunch adt3-userdebug; make checkbuild
Bug: 144863396
Change-Id: I9e5943539dbff6fd20d40679619cc25fd155d3f5