Don't use audio_config::frame_count to size buffer
adev_get_input_buffer_size is called with an audio_config_t with values
filled in for .sample_rate, .channel_mask, and .format. The Amlogic HAL
uses these to calculate the size of an audio buffer.

However, it also used the .frame_count. On most builds this is
initialized to zero, but there are some builds where it is uninitialized
and can be a large negative or positive number. The intended behavior
for a HAL is to just ignore this field since AudioFlinger will make its
own calculations.

Badly-behaving code shows output like this (from logcat). The calling
app crashes with SIGABRT.

    adev_get_input_buffer_size: exit: buffer_size = 3759769792

AudioFlinger fix in Change-Id: I41714cb2058a6fe7a6a9f55b1fd7c8d862d69351

Test: atest AudioRecordTest
Test: manual (open assistant app and speak)
Bug: 146119742
Change-Id: I59a9ba35248ecd3fe0e7e703ea5d8d5852d8a9b1