wifi: rewrite a new wifi HIDL service [2/2]
PD#SWPL-31274

Problem:
There is a wifi driver under hardware/broadcom/,
But this are some bugs in that driver.
So we changed it as other wifi module. But still
libwifi-hal still need link libcrypto.
But we can not change framework code.

Solution:
move wifi HIDL service to hardware/amlogic
and add a libwifi-hal-aml lib

Verify:
newton

Change-Id: Ia71f46be28f0be14c5f766bcffa83fee1000d252
Signed-off-by: Tellen Yu <tellen.yu@amlogic.com>