audio: Atmos_indication_ddp-MP4 and Atmos_prompt_ddp-MP4 [1/2]
PD#SWPL-54218

Problem:
Atmos_indication_ddp-MP4 and Atmos_prompt_ddp-MP4 [1/2]

Solution:
When DAP is enabl and dap_surround_virtualizer is OFF, the Dolby Atmos
indicator is not displayed.And "-dap_surround_virtualizer 0,96" is
disable Dolby Atmos, when the Dolby-Atmos is coming, the TV prompts the
user to switch on Dolby Atmos for the best experience.

Verify:
on Askey's SBR, these two cases are passed.

Change-Id: I5e63e8eac62217136a3a3db376688aabcd38c389
Signed-off-by: Xiushan Lu <xiushan.lu@amlogic.com>