audio: fix for plist passthrough drop [1/1]
PD#OTT-22257

Problem:
there is probabilistic drop in plist passthrough.

Solution:
the reason of this issue is the bypass latency is too low.
it leads to latency frames always is 0, as it can't adjust pst value.
so decrease the bypass latency and modify the logic latency frames can't
less than 0.

Verify:
S4D-MiBox

Signed-off-by: chuntian.miao <chuntian.miao@amlogic.com>
Change-Id: I0730a49c65dff2eb02347259b148de5b8df72759