health: fix build break with health hal aidl V2
When health aidl_interface version updraged, the libraries are
depending on V1 and libbatterymonitor or libhealthd_charger_ui at
the same time. It will cause multiple versions depending if
libbatterymonitor or libhealthd_charger_ui are depending on new
version.

Depend on the same version of libbatterymonitor, libhealthd_charger_ui
and health aidl_interface V1 to fix the problem.

Bug: 251425963
Change-Id: I0d95bde2fc3b1935a42e8d78a261eecd1294311e
Signed-off-by: Jack Wu <wjack@google.com>